some more password.c DO_RAY_IN_ZONE

+++ include/collision/block_22C84.h
@@ -27,7 +27,7 @@ void GET_RAY_ZDC(Obj *ray, s16 *x, s16 *y, s16 *w, s16 *h);
 void GET_BB1_ZDCs(Obj *obj, s16 *x_1, s16 *y_1, s16 *w_1, s16 *h_1, s16 *x_2, s16 *y_2, s16 *w_2, s16 *h_2);
 s32 myRand(s16 max_incl); /* ret: s32 (based on DO_MAIN_LOOP?) */
 void calc_obj_dir(Obj *obj);
-s32 OBJ_IN_ZONE(Obj *obj); /* ret: s32 or s16 (based on what, handle_GO_TEST?) */
+s16 OBJ_IN_ZONE(Obj *obj);
 void calc_obj_pos(Obj *obj);
 void makeUturn(Obj *obj);
 u16 BTYP(s16 x, s16 y);
